Service Layer Completed

+++ src/main/java/com/medkart/swiftmart/dto/OrderItemRequest.java
@@ -6,8 +6,8 @@ import lombok.Data;
 @Data
 public class OrderItemRequest {
 
-    private int productId;
-    private int quantity;
+    private Long productId;
+    private Long quantity;
 
 }
